terraform/full_environment/files/analytics_transfer_function/requirements.txt
functions_framework
google-cloud-bigquery
